create shared type
